pyproject.toml
[tool.poetry]
name = "panopticron"
version = "0.1.0"
description = "cron job to receive email notifications on daily gh activities of a particular user, for learning purposes."
authors = ["Adnene Guessoum <adnen.guessoum@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
flake8 = "^5.0.4"
pylint = "^2.15.5"
black = "^22.10.0"
mypy = "^0.982"
pytest = "^7.2.0"
flake8-pyproject = "^1.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.flake8]
max-line-length = 100

[tool.black]
max-line-length = 100
